Show trade polices as options in price table listing
Trade policies (AKA sales channels) also act as pricing tables, but only
the main one is included in the priceTables query. Introduce an
additional query to fetch the data for those and include it in the
display. This required a bit of tweaking of the UI logic since trade
policies have a display name plus a numerical id vs just a string name
for the regular price channels.